Prepare for next release (0.16.0)
Updates package version to 0.16.0.

Deprecates `dt_fst()` function and replaces it with `dt_parquet()`. The `dt_fst()` function is now defunct and throws an error when called.

Turn soft to warn deprecation in `add_adr()`, `add_drug()`, `get_drecno()` and `get_llt_smq()` and
warn to stop for `dt_fst()`, with relevant changes in its test and example